refactor(front): move skills-import request DTOs out of lib/api into types/api
Fifteenth batch: the internal request DTOs of the GitHub skills-import endpoint.

- Moved ImportSkillsRequestBodySchema + ImportSkillsRequestBody to
  types/api/skills/detection/github/import_skills.ts; dropped the now-unused zod
  import from the lib file.
- ImportSkillsResponseBody intentionally stays in lib/api (deferred): a "full"
  move requires porting SkillType -> SkillSchema -> the divergent
  MCPServerViewSchema into the published SDK, plus an SDK build/publish — its own
  project, tracked separately.
- 1 consumer (internal skills/import route) split: response body + function from
  @app/lib, request body + schema from @app/types.
