+++ front-api/routes/w/[wId]/skills/import/index.ts
@@ -1,12 +1,8 @@
-import type {
-  ImportSkillsRequestBody,
-  ImportSkillsResponseBody,
-} from "@app/lib/api/skills/detection/github/import_skills";
-import {
-  ImportSkillsRequestBodySchema,
-  importSkillsFromGitHub,
-} from "@app/lib/api/skills/detection/github/import_skills";
+import type { ImportSkillsResponseBody } from "@app/lib/api/skills/detection/github/import_skills";
+import { importSkillsFromGitHub } from "@app/lib/api/skills/detection/github/import_skills";
 import logger from "@app/logger/logger";
+import type { ImportSkillsRequestBody } from "@app/types/api/skills/detection/github/import_skills";
+import { ImportSkillsRequestBodySchema } from "@app/types/api/skills/detection/github/import_skills";
 import { assertNever } from "@app/types/shared/utils/assert_never";
 import { workspaceApp } from "@front-api/middlewares/ctx";
 import { ensureIsBuilder } from "@front-api/middlewares/ensure_role";
