+++ front/lib/api/skills/detection/github/import_skills.ts
@@ -27,16 +27,6 @@ import { Err, Ok } from "@app/types/shared/result";
 import { removeNulls } from "@app/types/shared/utils/general";
 import type { Octokit } from "@octokit/core";
 import path from "path";
-import { z } from "zod";
-
-export const ImportSkillsRequestBodySchema = z.object({
-  repoUrl: z.string(),
-  names: z.array(z.string()),
-});
-
-export type ImportSkillsRequestBody = z.infer<
-  typeof ImportSkillsRequestBodySchema
->;
 
 export type ImportSkillsResponseBody = {
   imported: SkillType[];
