+++ front/types/api/skills/detection/github/import_skills.ts
@@ -1,0 +1,10 @@
+import { z } from "zod";
+
+export const ImportSkillsRequestBodySchema = z.object({
+  repoUrl: z.string(),
+  names: z.array(z.string()),
+});
+
+export type ImportSkillsRequestBody = z.infer<
+  typeof ImportSkillsRequestBodySchema
+>;
